feat(rust): use already installed server if found

+++ lua/lspinstall/servers/rust.lua
@@ -4,26 +4,30 @@ config.default_config.cmd[1] = "./rust-analyzer"
 return vim.tbl_extend('error', config, {
   -- adjusted from https://github.com/mattn/vim-lsp-settings/blob/master/installer/install-rust-analyzer.sh
   install_script = [[
-  os=$(uname -s | tr "[:upper:]" "[:lower:]")
-  mchn=$(uname -m | tr "[:upper:]" "[:lower:]")
+  if ! [ -e "$PREFIX/bin/rust-analyzer"]; then
+    os=$(uname -s | tr "[:upper:]" "[:lower:]")
+    mchn=$(uname -m | tr "[:upper:]" "[:lower:]")
 
-  if [ $mchn = "arm64" ]; then
-    mchn="aarch64"
+    if [ $mchn = "arm64" ]; then
+      mchn="aarch64"
+    fi
+  
+    case $os in
+    linux)
+    platform="unknown-linux-gnu"
+    ;;
+    darwin)
+    platform="apple-darwin"
+    ;;
+    esac
+
+    curl -L -o "rust-analyzer-$mchn-$platform.gz" "https://github.com/rust-analyzer/rust-analyzer/releases/latest/download/rust-analyzer-$mchn-$platform.gz"
+    gzip -d rust-analyzer-$mchn-$platform.gz
+
+    mv rust-analyzer-$mchn-$platform rust-analyzer
+  else
+    ln "$PREFIX/bin/rust-analyzer" rust-analyzer
   fi
-
-  case $os in
-  linux)
-  platform="unknown-linux-gnu"
-  ;;
-  darwin)
-  platform="apple-darwin"
-  ;;
-  esac
-
-  curl -L -o "rust-analyzer-$mchn-$platform.gz" "https://github.com/rust-analyzer/rust-analyzer/releases/latest/download/rust-analyzer-$mchn-$platform.gz"
-  gzip -d rust-analyzer-$mchn-$platform.gz
-
-  mv rust-analyzer-$mchn-$platform rust-analyzer
 
   chmod +x rust-analyzer
   ]]
